+++ src/main/java/features/taskmanager/Event.java
@@ -1,4 +1,4 @@
-package features.event_manager;
+package features.taskmanager;
 import exceptions.InvalidArgumentException;
 import utils.Pair;
 import utils.TokenUtilities;
