+++ src/main/java/features/taskmanager/ToDo.java
@@ -1,4 +1,4 @@
-package features.event_manager;
+package features.taskmanager;
 
 import exceptions.InvalidArgumentException;
